Rename srml to paint (#32)
