+++ src/paint/mod.rs
